Cast trajectory areas to double
they used to be uints, which means that operations on them couldn't be negative (which was a problem when finding borders via diff(areas) )

+++ AP_get_probe_histology.m
@@ -214,8 +214,8 @@ switch eventdata.Key
                 use_trajectory_areas = ~isnan(trajectory_areas_uncut) & trajectory_areas_uncut > 1;
                 trajectory_areas = reshape(trajectory_areas_uncut(use_trajectory_areas),[],1);
                 
-                probe_ccf(curr_probe).trajectory_coords = trajectory_coords(use_trajectory_areas,:);
-                probe_ccf(curr_probe).trajectory_areas = trajectory_areas;
+                probe_ccf(curr_probe).trajectory_coords = double(trajectory_coords(use_trajectory_areas,:));
+                probe_ccf(curr_probe).trajectory_areas = double(trajectory_areas);
                 
             end
